requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pandas>=2.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
openpyxl>=3.1.0
Pillow>=10.0.0
pytesseract>=0.3.10
